// src/api/__generated__/wallet/model/otherUserInfoResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */
import type { FollowStatsResponse } from './followStatsResponse';
import type { OtherUserInfoResponseRelationStatus } from './otherUserInfoResponseRelationStatus';

/**
 * 他人用户信息
 */
export interface OtherUserInfoResponse {
  /** 用户ID */
  userId?: string;
  /** 用户昵称 */
  nickname?: string;
  /** 用户简介 */
  profile?: string;
  /** 用户头像URL */
  avatarUrl?: string;
  /** 用户状态：1-正常 0-禁用 */
  status?: string;
  followStats?: FollowStatsResponse;
  /** 当前登录用户对该用户的关系状态 */
  relationStatus?: OtherUserInfoResponseRelationStatus;
}