// src/api/__generated__/wallet/userwallet-user/userwallet-user.ts
export type otherUserInfoResponse200 = {
  data: OtherUserInfoResponse;
  status: 200;
};

export type otherUserInfoResponse400 = {
  data: ResultObject;
  status: 400;
};

export type otherUserInfoResponse405 = {
  data: ResultObject;
  status: 405;
};

export type otherUserInfoResponse500 = {
  data: ResultObject;
  status: 500;
};

export type otherUserInfoResponseSuccess = otherUserInfoResponse200 & {
  headers: Headers;
};
export type otherUserInfoResponseError = (
  | otherUserInfoResponse400
  | otherUserInfoResponse405
  | otherUserInfoResponse500
) & {
  headers: Headers;
};

export type otherUserInfoResponse =
  | otherUserInfoResponseSuccess
  | otherUserInfoResponseError;